rag-service/requirements.txt
langchain==0.1.14
langchain-community==0.0.30
qdrant-client==1.7.0
sentence-transformers==2.2.2
pypdf==3.17.4
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.0.1
